refactor: move `NoteQuantizationUnit` out to `music` package
