+++ src/main/java/site/overwrite/auditranscribe/io/data_files/data_encapsulators/SettingsData.java
@@ -20,7 +20,7 @@ package site.overwrite.auditranscribe.io.data_files.data_encapsulators;
 
 import site.overwrite.auditranscribe.audio.WindowFunction;
 import site.overwrite.auditranscribe.misc.Theme;
-import site.overwrite.auditranscribe.music.notes.NoteQuantizationUnit;
+import site.overwrite.auditranscribe.music.NoteUnit;
 import site.overwrite.auditranscribe.spectrogram.ColourScale;
 
 /**
@@ -38,7 +38,7 @@ public class SettingsData {
     // Transcription data
     public int colourScaleEnumOrdinal = ColourScale.VIRIDIS.ordinal();
     public int windowFunctionEnumOrdinal = WindowFunction.HANN_WINDOW.ordinal();
-    public int noteQuantizationUnitEnumOrdinal = NoteQuantizationUnit.THIRTY_SECOND_NOTE.ordinal();
+    public int noteQuantizationUnitEnumOrdinal = NoteUnit.THIRTY_SECOND_NOTE.ordinal();
 
     // Miscellaneous data
     public int themeEnumOrdinal = Theme.LIGHT_MODE.ordinal();
